New border styles
The new border style is more general and easier to read
* The row sep is a string, not a character
* Each row (top, text, mid, bot) has its own separator
* The separator for text lines is the pad string
  " " was the only pad string in the old style

A frame style is a special case of the new style,
where each element is a single character and thus the whole
style can easily be written as a list of 4 strings like
    (double
     "╔═╦╗"
     "║ ║║"
     "╠═╬╣"
     "╚═╩╝")

The old style is still supported, and everything should be
backward compatible.

Docs still need updating.