fix #12
The example in the docs now uses `'(#t #f ...)` to avoid any confusion.

`pattern-list->list` has an argument `#:truncate-ok?` which reduces the returned list to exactly
`n-elts` even when the original list as more elements than `n-elts`.

However, `(pattern-list->list '(#t #f))` remains equivalent to `(pattern-list->list '(#t #f #f ...))`.
That is, the minimum length of the list is 2 and an exception is thrown unless `#:truncate-ok? #t`
is passed.